+++ modules/optionals/datalake-buckets/tests/test_app.py
@@ -9,13 +9,14 @@ import pytest
 
 @pytest.fixture(scope="function")
 def stack_defaults():
-    os.environ["ADDF_DEPLOYMENT_NAME"] = "test-proj"
-    os.environ["ADDF_MODULE_NAME"] = "test-dep"
-    os.environ["ADDF_HASH"] = "hash"
+    os.environ["SEEDFARMER_PROJECT_NAME"] = "test-project"
+    os.environ["SEEDFARMER_DEPLOYMENT_NAME"] = "test-deployment"
+    os.environ["SEEDFARMER_MODULE_NAME"] = "test-module"
+    os.environ["SEEDFARMER_HASH"] = "hash"
     os.environ["CDK_DEFAULT_ACCOUNT"] = "111111111111"
     os.environ["CDK_DEFAULT_REGION"] = "us-east-1"
-    os.environ["ADDF_PARAMETER_ENCRYPTION_TYPE"] = "SSE"
-    os.environ["ADDF_PARAMETER_RETENTION_TYPE"] = "DESTROY"
+    os.environ["SEEDFARMER_PARAMETER_ENCRYPTION_TYPE"] = "SSE"
+    os.environ["SEEDFARMER_PARAMETER_RETENTION_TYPE"] = "DESTROY"
 
     # Unload the app import so that subsequent tests don't reuse
     if "app" in sys.modules:
@@ -27,39 +28,39 @@ def test_app(stack_defaults):
 
 
 def test_buckets_encryption_type(stack_defaults):
-    del os.environ["ADDF_PARAMETER_ENCRYPTION_TYPE"]
+    del os.environ["SEEDFARMER_PARAMETER_ENCRYPTION_TYPE"]
 
     with pytest.raises(Exception):
         import app  # noqa: F401
 
-        assert os.environ["ADDF_PARAMETER_ENCRYPTION_TYPE"] == "SSE"
+        assert os.environ["SEEDFARMER_PARAMETER_ENCRYPTION_TYPE"] == "SSE"
 
 
 def test_invalid_buckets_encryption_type(stack_defaults):
-    os.environ["ADDF_PARAMETER_ENCRYPTION_TYPE"] = "notvalid"
+    os.environ["SEEDFARMER_PARAMETER_ENCRYPTION_TYPE"] = "notvalid"
     with pytest.raises(Exception):
         import app  # noqa: F401
 
 
 def test_buckets_retention(stack_defaults):
-    del os.environ["ADDF_PARAMETER_RETENTION_TYPE"]
+    del os.environ["SEEDFARMER_PARAMETER_RETENTION_TYPE"]
 
     with pytest.raises(Exception):
         import app  # noqa: F401
 
-        assert os.environ["ADDF_PARAMETER_RETENTION_TYPE"] == "DESTROY"
+        assert os.environ["SEEDFARMER_PARAMETER_RETENTION_TYPE"] == "DESTROY"
 
 
 def test_invalid_retention_type(stack_defaults):
-    os.environ["ADDF_PARAMETER_RETENTION_TYPE"] = "notvalid"
+    os.environ["SEEDFARMER_PARAMETER_RETENTION_TYPE"] = "notvalid"
     with pytest.raises(Exception):
         import app  # noqa: F401
 
 
 def test_solution_description(stack_defaults):
-    os.environ["ADDF_PARAMETER_SOLUTION_ID"] = "SO123456"
-    os.environ["ADDF_PARAMETER_SOLUTION_NAME"] = "MY GREAT TEST"
-    os.environ["ADDF_PARAMETER_SOLUTION_VERSION"] = "v1.0.0"
+    os.environ["SEEDFARMER_PARAMETER_SOLUTION_ID"] = "SO123456"
+    os.environ["SEEDFARMER_PARAMETER_SOLUTION_NAME"] = "MY GREAT TEST"
+    os.environ["SEEDFARMER_PARAMETER_SOLUTION_VERSION"] = "v1.0.0"
 
     import app
 
@@ -68,9 +69,9 @@ def test_solution_description(stack_defaults):
 
 
 def test_solution_description_no_version(stack_defaults):
-    os.environ["ADDF_PARAMETER_SOLUTION_ID"] = "SO123456"
-    os.environ["ADDF_PARAMETER_SOLUTION_NAME"] = "MY GREAT TEST"
-    del os.environ["ADDF_PARAMETER_SOLUTION_VERSION"]
+    os.environ["SEEDFARMER_PARAMETER_SOLUTION_ID"] = "SO123456"
+    os.environ["SEEDFARMER_PARAMETER_SOLUTION_NAME"] = "MY GREAT TEST"
+    del os.environ["SEEDFARMER_PARAMETER_SOLUTION_VERSION"]
 
     import app